feat: add async wrapper for resolver
Added behind the async feature.

Additionally, add an `as_ptr` method to ngx::core::Pool.

This was originally an upstreaming of nginx-acme::net::resolver, but was
rewritten during this upstreaming process see:
https://github.com/nginx/nginx-acme/blob/main/src/net/resolver.rs

There are no tests for this functionality in the ngx-rust repo at this
time. This PR has been tested by manually using a test suite in a
private nginx repo. Tests will be added in this repo at some point in
the future.

Co-authored-by: Evgeny Shirykalov <e.shirykalov@f5.com>
Co-authored-by: Aleksei Bavshin <a.bavshin@f5.com>